feat: update makefile (#238)
* feat: update Makefile to use HUGO_CMD for all commands

This commit modifies the Makefile to standardize the use of the HUGO_CMD variable for executing Hugo commands.

## Background
The previous implementation directly called Hugo commands, which could lead to inconsistencies in command execution.

## Changes
- Replaced direct Hugo commands with the HUGO_CMD variable in the update, build, server, post, news, publication, and event targets.

* feat: add publication-thumbnail target to Makefile and update README

## Background
This commit introduces a new Makefile target for generating a thumbnail image for publications. It enhances the build process by allowing users to create a featured image directly from a PDF file.

## Changes
- Added `publication-thumbnail` target in the Makefile that utilizes ImageMagick to create a thumbnail from a specified PDF.
- Updated the README to reflect the new command for creating a publication thumbnail, replacing the previous command with a more streamlined `make` command.

* feat: update go.mod and go.sum with new dependencies

## Background
This commit adds new dependencies to the project, ensuring that the necessary modules are included for proper functionality.

## Changes
- Updated `go.mod` to include:
  - `blox-bootstrap` version `v5.9.7`
  - `blox-core` version `v0.3.1`
  - `blox-plugin-netlify` version `v1.1.1`
  - `blox-plugin-reveal` version `v1.1.2`
  - `blox-seo` version `v0.2.3`

- Updated `go.sum` to reflect the new dependencies and their respective checksums.

* fix: correct pagination configuration in hugo.yaml

## Background
The pagination configuration in the hugo.yaml file was incorrectly defined, which could lead to issues in paginating content on the website.

## Changes
- Changed `paginate: 10` to a structured `pagination:` block with `pagerSize: 10`.

* fix: update ignoreFiles in hugo.yaml

## Background
The ignoreFiles configuration in the hugo.yaml file needed to be updated to include an additional pattern for template configuration files.

## Changes
- Added `template\.config\.yaml$` to the ignoreFiles list to ensure these files are not processed by Hugo.