feat: update for FOUND workshop (#257)
* feat: add news abount accepted workshop proposal for ICCV 2025

This commit introduces a new page detailing the acceptance of our workshop proposal on Foundation Data at ICCV 2025.

## Background
The workshop aims to emphasize the importance of data in developing foundation models, focusing on robust dataset construction and industry applications.

## Changes
- Created a new markdown file for the workshop proposal.
- Included metadata such as title, authors, tags, and categories.
- Added a link to the workshop's website for further information.

* feat: add ICCV 2025 workshop announcement and details

This commit introduces the announcement for the workshop titled "FOUND: Foundation Data for Industrial Tech Transfer" accepted at ICCV 2025.

## Background
The workshop aims to address the critical role of data in building foundation models, focusing on robust dataset construction and industry-driven applications.

## Changes
- Added a new markdown file for the workshop announcement.
- Included detailed information about the workshop, topics of interest, and the significance of foundation data in AI research and applications.
- Added a featured image for the workshop.

* feat: update workshop proposal title for clarity

## Background
The title of the workshop proposal announcement needed to be more precise to reflect the acceptance status.

## Changes
- Changed the title from "Our Workshop Proposal on Foundation Data Accepted at ICCV 2025" to "Accepted our Workshop Proposal on Foundation Data to ICCV 2025".